fw-4200, add mixin for site content viewsets
