+++ firstvoices/backend/views/base_views.py
@@ -1,6 +1,9 @@
+from django.core.exceptions import PermissionDenied
+from django.http import Http404
 from rest_framework.response import Response
 from rules.contrib.rest_framework import AutoPermissionViewSetMixin
 
+from backend.models import Site
 from backend.predicates import utils
 
 
@@ -36,3 +39,21 @@ class FVPermissionViewSetMixin(AutoPermissionViewSetMixin):
             queryset, many=True, context={"request": request}
         )
         return Response(serializer.data)
+
+
+class SiteContentViewSetMixin:
+    """
+    Provides common methods for handling site content, usually for data models that use the BaseSiteContentModel.
+    """
+
+    def get_validated_site(self):
+        site_slug = self.kwargs["site_slug"]
+        site = Site.objects.filter(slug=site_slug)
+
+        if site.count() == 0:
+            raise Http404
+
+        if utils.filter_by_viewable(self.request.user, site).count() == 0:
+            raise PermissionDenied
+
+        return site
