+++ firstvoices/backend/views/dictionary_views.py
@@ -1,12 +1,9 @@
-from django.core.exceptions import PermissionDenied
-from django.http import Http404
 from drf_spectacular.utils import OpenApiResponse, extend_schema, extend_schema_view
 from rest_framework.viewsets import ModelViewSet
 
-from backend.models import DictionaryEntry, Site
-from backend.predicates import utils
+from backend.models import DictionaryEntry
 from backend.serializers.dictionary_serializers import DictionaryEntryDetailSerializer
-from backend.views.base_views import FVPermissionViewSetMixin
+from backend.views.base_views import FVPermissionViewSetMixin, SiteContentViewSetMixin
 
 
 @extend_schema_view(
@@ -27,25 +24,15 @@ from backend.views.base_views import FVPermissionViewSetMixin
         },
     ),
 )
-class DictionaryViewSet(FVPermissionViewSetMixin, ModelViewSet):
+class DictionaryViewSet(
+    FVPermissionViewSetMixin, SiteContentViewSetMixin, ModelViewSet
+):
     """
     Dictionary entry information.
     """
 
     http_method_names = ["get"]
     serializer_class = DictionaryEntryDetailSerializer
-
-    def get_validated_site(self):
-        site_slug = self.kwargs["site_slug"]
-        site = Site.objects.filter(slug=site_slug)
-
-        if site.count() == 0:
-            raise Http404
-
-        if utils.filter_by_viewable(self.request.user, site).count() == 0:
-            raise PermissionDenied
-
-        return site
 
     def get_queryset(self):
         site = self.get_validated_site()
